requirements.txt
flask
opencv-python
numpy
tensorflow
pillow
flask-cors